+++ src/Felicity.IntegrationTests/ToManyRelationshipPatchResourceAndSelf.fs
@@ -872,6 +872,32 @@ let tests1 =
       test <@ json |> hasNoPath "errors[1]" @>
     }
 
+    testJob "Returns 400 for each null resource identifier" {
+      let db = Db ()
+      let! response =
+        Request.patch (Ctx.WithDb db) "/parents/p1"
+        |> Request.bodySerialized
+            {|data =
+                {|``type`` = "parent1"
+                  id = "p1"
+                  relationships =
+                    {|children =
+                        {| data = [ null; box {| ``type`` = "child2"; id = "foo" |}; null ] |}
+                    |}
+                |}
+            |}
+        |> getResponse
+      response |> testStatusCode 400
+      let! json = response |> Response.readBodyAsString
+      test <@ json |> getPath "errors[0].status" = "400" @>
+      test <@ json |> getPath "errors[0].detail" = "Array 'data' may not have null items" @>
+      test <@ json |> getPath "errors[0].source.pointer" = "/data/relationships/children/data/0" @>
+      test <@ json |> getPath "errors[1].status" = "400" @>
+      test <@ json |> getPath "errors[1].detail" = "Array 'data' may not have null items" @>
+      test <@ json |> getPath "errors[1].source.pointer" = "/data/relationships/children/data/2" @>
+      test <@ json |> hasNoPath "errors[2]" @>
+    }
+
     testJob "Returns 400 for each relationship data identifier that has missing type" {
       let db = Db ()
       let! response =
@@ -1535,6 +1561,23 @@ let tests2 =
       test <@ json |> getPath "errors[0].detail" = "Member 'data' may not be null" @>
       test <@ json |> getPath "errors[0].source.pointer" = "/data" @>
       test <@ json |> hasNoPath "errors[1]" @>
+    }
+
+    testJob "Returns 400 for each null resource identifier" {
+      let db = Db ()
+      let! response =
+        Request.patch (Ctx.WithDb db) "/parents/p1/relationships/children"
+        |> Request.bodySerialized {| data = [ null; box {| ``type`` = "child2"; id = "foo" |}; null ] |}
+        |> getResponse
+      response |> testStatusCode 400
+      let! json = response |> Response.readBodyAsString
+      test <@ json |> getPath "errors[0].status" = "400" @>
+      test <@ json |> getPath "errors[0].detail" = "Array 'data' may not have null items" @>
+      test <@ json |> getPath "errors[0].source.pointer" = "/data/0" @>
+      test <@ json |> getPath "errors[1].status" = "400" @>
+      test <@ json |> getPath "errors[1].detail" = "Array 'data' may not have null items" @>
+      test <@ json |> getPath "errors[1].source.pointer" = "/data/2" @>
+      test <@ json |> hasNoPath "errors[2]" @>
     }
 
     testJob "Returns 400 for each missing type" {