+++ RELEASE_NOTES.md
@@ -1,6 +1,10 @@
 Release notes
 ==============
 
+### Unreleased
+
+* Fixed inadvertently returning a 500 error instead of a 400 error when a request has `null` items in a to-many relationship's `data` array or in a resource identifier collection document's primary `data` array. 
+
 ### 0.21.1 (2022-10-09)
 
 * Updated FSharp.SystemTextJson from 0.19.13 to 1.0.6. There is no change in behavior in Felicity, but note that this comes with a breaking change if you use it in other places: When deserializing missing fields of type `option` or `voption`, an error will now be returned instead of deserializing to `null`. Either wrap such fields in `Skippable`, or enable the option `IgnoreNullValues = true`. You can find [more details here](https://tarmil.fr/article/2022/9/25/systemtextjson-v1.0).